requirements.txt
# ===============================================
# ScheduleBot Production Dependencies
# Python >= 3.11, < 3.13
# ===============================================

# === Telegram Bot Framework ===
aiogram>=3.15.0,<3.23.0

# === Web Framework ===
fastapi>=0.115.0,<0.123.0

# === Database ===
aiosqlite>=0.20.0

# === Task Scheduler ===
apscheduler>=3.10.4,<4.0.0

# === Data Validation ===
pydantic>=2.7.0,<2.11.0
pydantic-settings>=2.3.0,<3.0.0

# === Template Engine ===
jinja2>=3.1.3,<4.0.0

# === ASGI Server ===
uvicorn>=0.30.0,<0.32.0

# === Utilities ===
python-dotenv>=1.0.0
anyio>=4.4.0,<5.0.0
aiofiles>=23.0.0,<24.2.0
typing-extensions>=4.10.0

# === Sessions ===
itsdangerous>=2.1.0

# === HTTP Forms ===
python-multipart>=0.0.9
